+++ qupath-core/src/main/java/qupath/lib/io/UriResource.java
@@ -55,4 +55,6 @@ public interface UriResource {
 	 */
 	public boolean updateUris(Map<URI, URI>  replacements) throws IOException;
 
+	
+	
 }